Перенес main в место запуска __init__
